+++ npm/vue/examples/cli2/plugins.js
@@ -1,8 +1,0 @@
-/// <reference types="cypress" />
-const preprocessor = require('../../dist/plugins/webpack')
-
-module.exports = (on, config) => {
-  preprocessor(on, config, require('@vue/cli-service/webpack.config'))
-
-  return config
-}