+++ packages/server/lib/browsers/index.js
@@ -11,34 +11,35 @@ const isBrowserFamily = check.oneOf(['chromium', 'firefox'])
 
 let instance = null
 
-const kill = function (unbind) {
-  // cleanup our running browser
-  // instance
+const kill = function (unbind, isProcessExit) {
+  // Clean up the instance when the browser is closed
   if (!instance) {
+    debug('browsers.kill called with no active instance')
+
     return Promise.resolve()
   }
 
-  return new Promise((resolve) => {
-    if (unbind) {
-      instance.removeAllListeners()
-    }
+  const _instance = instance
 
-    instance.once('exit', (...args) => {
+  instance = null
+
+  if (unbind) {
+    _instance.removeAllListeners()
+  }
+
+  return new Promise((resolve) => {
+    _instance.once('exit', () => {
       debug('browser process killed')
 
-      return resolve.apply(null, args)
+      resolve()
     })
 
     debug('killing browser process')
 
-    instance.kill()
+    _instance.isProcessExit = isProcessExit
 
-    return cleanup()
+    _instance.kill()
   })
-}
-
-const cleanup = () => {
-  return instance = null
 }
 
 const getBrowserLauncher = function (browser) {
@@ -141,7 +142,7 @@ const throwBrowserNotFound = function (browserName, browsers = []) {
   return errors.throw('BROWSER_NOT_FOUND_BY_NAME', browserName, names)
 }
 
-process.once('exit', kill)
+process.once('exit', () => kill(true, true))
 
 module.exports = {
   ensureAndGetByNameOrPath,
@@ -151,8 +152,6 @@ module.exports = {
   removeOldProfiles: utils.removeOldProfiles,
 
   get: utils.getBrowsers,
-
-  launch: utils.launch,
 
   close: kill,
 
@@ -211,8 +210,7 @@ module.exports = {
         // enable the browser to configure the interface
         instance.once('exit', () => {
           options.onBrowserClose()
-
-          return cleanup()
+          instance = null
         })
 
         // TODO: instead of waiting an arbitrary
@@ -235,5 +233,4 @@ module.exports = {
       })
     })
   },
-
 }