+++ packages/server/lib/cypress.js
@@ -60,10 +60,6 @@ const exitErr = (err) => {
   })
 }
 
-const isComponentTesting = (options) => {
-  return options.testingType === 'component'
-}
-
 module.exports = {
   isCurrentlyRunningElectron () {
     return require('./util/electron-app').isRunning()
@@ -78,20 +74,14 @@ module.exports = {
       // that means we're already running in electron
       // like in production and we shouldn't spawn a new
       // process
-      if (isComponentTesting(options) || this.isCurrentlyRunningElectron()) {
+      if (this.isCurrentlyRunningElectron()) {
         // if we weren't invoked from the CLI
         // then display a warning to the user
         if (!options.invokedFromCli) {
           warning('INVOKED_BINARY_OUTSIDE_NPM_MODULE')
         }
 
-        // just run the gui code directly here
-        // and pass our options directly to main
-        if (isComponentTesting(options)) {
-          debug(`skipping running Electron when in ${mode} mode`)
-        } else {
-          debug('running Electron currently')
-        }
+        debug('running Electron currently')
 
         return require('./modes')(mode, options)
       }
@@ -148,11 +138,6 @@ module.exports = {
     }
 
     debug('from argv %o got options %o', argv, options)
-
-    // Allow for Cypress to test locally, but do not allow users to access component testing
-    if (options.componentTesting && !process.env.CYPRESS_INTERNAL_ENV) {
-      throw new Error('Component testing mode is not implemented. But coming 🥳.')
-    }
 
     if (options.headless) {
       // --headless is same as --headed false