+++ packages/server/lib/experiments.ts
@@ -46,13 +46,13 @@ interface StringValues {
  * @example
   ```
   {
-    experimentalComponentTesting: 'Allows mounting and testing framework-specific components'
+    experimentalFetchPolyfill: 'Polyfills `window.fetch` to enable Network spying and stubbing.'
   }
   ```
 */
 const _summaries: StringValues = {
-  experimentalComponentTesting: 'Framework-specific component testing, uses `componentFolder` to load component specs.',
   experimentalFetchPolyfill: 'Polyfills `window.fetch` to enable Network spying and stubbing.',
+  experimentalInteractiveRunEvents: 'Allows listening to the `before:run`, `after:run`, `before:spec`, and `after:spec` events in the plugins file during interactive mode.',
   experimentalSourceRewriting: 'Enables AST-based JS/HTML rewriting. This may fix issues caused by the existing regex-based JS/HTML replacement algorithm.',
   experimentalStudio: 'Generate and save commands directly to your test suite by interacting with your app as an end user would.',
 }
@@ -63,13 +63,13 @@ const _summaries: StringValues = {
  * @example
   ```
   {
-    experimentalComponentTesting: 'Component Testing'
+    experimentalFetchPolyfill: 'Fetch polyfill'
   }
   ```
 */
 const _names: StringValues = {
-  experimentalComponentTesting: 'Component Testing',
   experimentalFetchPolyfill: 'Fetch polyfill',
+  experimentalInteractiveRunEvents: 'Interactive Mode Run Events',
   experimentalSourceRewriting: 'Improved source rewriting',
   experimentalStudio: 'Studio',
 }