+++ packages/web-config/webpack.config.base.ts
@@ -30,6 +30,31 @@ const evalDevToolPlugin = new webpack.EvalDevToolModulePlugin({
 
 evalDevToolPlugin.evalDevToolPlugin = true
 
+const optimization = {
+  usedExports: true,
+  providedExports: true,
+  sideEffects: true,
+  namedChunks: true,
+  namedModules: true,
+  removeAvailableModules: true,
+  mergeDuplicateChunks: true,
+  flagIncludedChunks: true,
+  removeEmptyChunks: true,
+}
+
+const stats = {
+  errors: true,
+  warningsFilter: /node_modules\/mocha\/lib\/mocha.js/,
+  warnings: true,
+  all: false,
+  builtAt: true,
+  colors: true,
+  modules: true,
+  maxModules: 20,
+  excludeModules: /(main|test-entry).scss/,
+  timings: true,
+}
+
 function makeSassLoaders ({ modules }): RuleSetRule {
   const exclude = [/node_modules/]
 
@@ -72,7 +97,7 @@ function makeSassLoaders ({ modules }): RuleSetRule {
   }
 }
 
-const getCommonConfig = () => {
+export const getCommonConfig = () => {
   const commonConfig: Configuration = {
     mode: 'none',
     node: {
@@ -86,18 +111,8 @@ const getCommonConfig = () => {
       extensions: ['.ts', '.js', '.jsx', '.tsx', '.scss', '.json'],
     },
 
-    stats: {
-      errors: true,
-      warningsFilter: /node_modules\/mocha\/lib\/mocha.js/,
-      warnings: true,
-      all: false,
-      builtAt: true,
-      colors: true,
-      modules: true,
-      maxModules: 20,
-      excludeModules: /(main|test-entry).scss/,
-      timings: true,
-    },
+    stats,
+    optimization,
 
     module: {
       rules: [
@@ -113,7 +128,10 @@ const getCommonConfig = () => {
                 [require.resolve('@babel/plugin-proposal-class-properties'), { loose: true }],
               ],
               presets: [
-                [require.resolve('@babel/preset-env'), { targets: { 'chrome': 63 } }],
+                // the chrome version should be synced with
+                // npm/webpack-batteries-included-preprocessor/index.js and
+                // packages/server/lib/browsers/chrome.ts
+                [require.resolve('@babel/preset-env'), { targets: { 'chrome': '64' } }],
                 require.resolve('@babel/preset-react'),
                 [require.resolve('@babel/preset-typescript'), { allowNamespaces: true }],
               ],
@@ -164,18 +182,6 @@ const getCommonConfig = () => {
       ],
     },
 
-    optimization: {
-      usedExports: true,
-      providedExports: true,
-      sideEffects: true,
-      namedChunks: true,
-      namedModules: true,
-      removeAvailableModules: true,
-      mergeDuplicateChunks: true,
-      flagIncludedChunks: true,
-      removeEmptyChunks: true,
-    },
-
     plugins: [
       new CleanWebpackPlugin({ cleanStaleWebpackAssets: false }),
       new MiniCSSExtractWebpackPlugin(),
@@ -208,6 +214,38 @@ const getCommonConfig = () => {
   return commonConfig
 }
 
-export default getCommonConfig
+// eslint-disable-next-line @cypress/dev/arrow-body-multiline-braces
+export const getSimpleConfig = () => ({
+  resolve: {
+    extensions: ['.js'],
+  },
+
+  stats,
+  optimization,
+
+  cache: true,
+
+  module: {
+    rules: [
+      {
+        test: /\.(js)$/,
+        exclude: /node_modules/,
+        use: {
+          loader: require.resolve('babel-loader'),
+          options: {
+            presets: [
+              [require.resolve('@babel/preset-env'), { targets: { 'chrome': 63 } }],
+            ],
+            babelrc: false,
+          },
+        },
+      },
+    ],
+  },
+
+  plugins: [
+    new CleanWebpackPlugin({ cleanStaleWebpackAssets: false }),
+  ],
+})
 
 export { HtmlWebpackPlugin }